unidad1/actividad3/proyecto final Valeria y Luis/backend/requirements.txt
Django==5.2.8
djangorestframework==3.14.0
requests==2.31.0
